Create post-receive git web hook
This will deploy the master branch of the git repository to the defined web directory. It is designed to deploy Laravel applications so that it will clear the config and route caches and run a force migrations as it deploys. It may need to use sudo for some/all of the commands depending on your set-up. The initial PATH stuff may not be needed depending on your set-up as well.

Note: this is as-is and use at your own risk and all that stuff.